Trait ic_cow_state::CowMemoryManager

pub trait CowMemoryManager {
    fn get_map(&self) -> MappedStateImpl;

    fn get_map_for_snapshot(
        &self, 
        _round_to_use: u64
    ) -> Result<MappedStateImpl, CowError>;

    fn create_snapshot(&self, end_round: u64);

    fn checkpoint(&self);

    fn upgrade(&self);

    fn is_valid(&self) -> bool;

    fn remove_states_below(&self, round: u64);
}

Required methods

fn get_map(&self) -> MappedStateImpl

get_map returns a MappedState representing "current" mapped state of the canister (heap and globals for time being). This state can be used for example during canister execution and can be freely mutated. "current_state" can be updated with mutations by calling "soft_commmit".

fn get_map_for_snapshot(
&self,
_round_to_use: u64
) -> Result<MappedStateImpl, CowError>

get_map_for_snapshot returns a MappedState representing in memory mapped state of canister (heap and globals for time being) at the end of a specific round. This state, although can be freely mutated, the mutations cannot be made part round state using "soft_commit". MappedState returned by get_map_for_snapshot can be used for query type canister operations.

fn create_snapshot(&self, end_round: u64)

create_snapshot creates a snapshot of all soft_committed mutations to canister state the last snapshot

fn checkpoint(&self)

checkpoint primarily ensures slot_mgr's internal metadata is flushed to disk Collapsing older rounds would be added here later.

fn upgrade(&self)

Reset's canister's "current" state to "vanilla" initial state.

fn is_valid(&self) -> bool

fn remove_states_below(&self, round: u64)

Loading content...