Trait ic_cow_state::MappedState

pub trait MappedState: Send {
    fn get_heap_base(&self) -> *mut u8;

    fn get_heap_len(&self) -> usize;

    fn make_heap_accessible(&self);

    fn get_globals(&self) -> &[u8];

    fn copy_from_heap(&self, _offset: u64, _len: u64) -> &[u8];

    fn update_globals(&self, _encoded_globals: &[u8]) { ... }

    fn soft_commit(&self, _pages: &[u64]) { ... }

    fn update_heap_page(&self, _page_idx: u64, _bytes: &[u8]) { ... }

    fn copy_to_heap(&self, _offset: u64, _bytes: &[u8]) -> Vec<u64> { ... }

    fn clear(&self) { ... }
}

Required methods

fn get_heap_base(&self) -> *mut u8

Returns the base of memory region where canister heap begins.

fn get_heap_len(&self) -> usize

Returns the max heap len when the heap was first initialized.

fn make_heap_accessible(&self)

Resets the permission on the heap based on the type

fn get_globals(&self) -> &[u8]

Returns slice of globals memory associated with this mapped_state.

fn copy_from_heap(&self, _offset: u64, _len: u64) -> &[u8]

Loading content...
Provided methods

fn update_globals(&self, _encoded_globals: &[u8])

Update the globals memory with new globals.

fn soft_commit(&self, _pages: &[u64])

soft_commit ensures that all modified pages (mutations) represented by pages slice become part of the "current" state. This means even after MappedState object is dropped, mutations will be reflected in subsequent MappedState object and can eventually become part of some round. Mutations not soft_committed have a lifetime of MappedState object and will disappear once MappedState object is dropped.

fn update_heap_page(&self, _page_idx: u64, _bytes: &[u8])

Using update_heap_page, individual heap pages can be modified. This is useful during canister installation to patch up initial pages. Once the heap is modified, soft_commit is necessary to ensure that the modifications become part of "current" state.

fn copy_to_heap(&self, _offset: u64, _bytes: &[u8]) -> Vec<u64>

fn clear(&self)

Loading content...